refactor(auth-core): make code more DRY
