+++ plugins/auth-core/src/core/create-token-payload-interface.ts
@@ -5,6 +5,7 @@ import { builders, namedTypes } from "ast-types";
 import { getUserIdType } from "../util/get-user-id-type";
 import { join } from "path";
 import { templatesPath } from "../constants";
+import { idTypeTSOptions } from "../util/idTypeMapper";
 
 const templatePath = join(
   templatesPath,
@@ -18,24 +19,10 @@ export async function createTokenPayloadInterface(
   const authDir = `${serverDirectories.authDirectory}`;
   const template = await readFile(templatePath);
   const idType = getUserIdType(dsgContext);
-  const templateMapping = prepareTemplateMapping(idType);
+  const templateMapping = { ID_TYPE: idTypeTSOptions[idType] };
   const filePath = `${authDir}/ITokenService.ts`;
   interpolate(template, templateMapping);
   removeTSClassDeclares(template);
 
   return { code: print(template).code, path: filePath };
-}
-
-function prepareTemplateMapping(idType: types.Id["idType"]) {
-  /* eslint-disable @typescript-eslint/naming-convention */
-  const idTypeTSOptions: {
-    [key in types.Id["idType"]]: namedTypes.Identifier;
-  } = {
-    AUTO_INCREMENT: builders.identifier("number"),
-    AUTO_INCREMENT_BIG_INT: builders.identifier("number"),
-    UUID: builders.identifier("string"),
-    CUID: builders.identifier("string"),
-  };
-
-  return { ID_TYPE: idTypeTSOptions[idType] };
 }
